Move the JetCorrectorParametersHelper class out from inside the JetCorrectorParameters class. Also untemplate the JetCorrectorParametersHelper class and set a maximum number of binned variables to 3.
