+++ CondFormats/JetMETObjects/src/classes_def.xml
@@ -3,10 +3,10 @@
   <class name="FactorizedJetCorrector"/>
   <class name="JetCorrectionUncertainty"/>
   <class name="JetCorrectorParameters">
-    <field name="helperTuple"             transient="true"/>
+    <field name="helper"                  transient="true"/>
     <field name="MAX_SIZE_DIMENSIONALITY" transient="true"/>
   </class>
-  <ioread sourceClass="JetCorrectorParameters" targetClass="JetCorrectorParameters" version="[0-]" source="std::vector<JetCorrectorParameters::Record> mRecords" target="">
+  <ioread sourceClass="JetCorrectorParameters" targetClass="JetCorrectorParameters" version="[0-]" source="JetCorrectorParameters::Definitions mDefinitions; std::vector<JetCorrectorParameters::Record> mRecords;" target="">
     <![CDATA[
          newObj->init(); 
     ]]>
